Extract model coefficients from a fitted textmodel_ca object

coef() extract model coefficients from a fitted textmodel_ca object. coefficients() is an alias.

# S3 method for textmodel_ca
coef(object, doc_dim = 1, feat_dim = 1, ...)

coefficients.textmodel_ca(object, doc_dim = 1, feat_dim = 1, ...)

Arguments

object

a fitted textmodel_ca object

doc_dim, feat_dim

the document and feature dimension scores to be extracted

...

unused

Value

a list containing numeric vectors of feature and document coordinates. Includes NA vectors of standard errors for consistency with other models' coefficient outputs, and for the possibility of having these computed in the future.

coef_feature column coordinates of the features

coef_feature_se feature length vector of NA values

coef_document row coordinates of the documents

coef_document_se document length vector of NA values