Movie reviews with polarity from Pang and Lee (2004)

A corpus object containing 2,000 movie reviews classified by positive or negative sentiment.

data_corpus_moviereviews

Format

The corpus includes the following document variables:

sentiment

factor indicating whether a review was manually classified as positive pos or negative neg.

id1

Character counting the position in the corpus.

id2

Random number for each review.

Source

https://www.cs.cornell.edu/people/pabo/movie-review-data/

Details

For more information, see cat(meta(data_corpus_moviereviews, "readme")).

References

Pang, B., Lee, L. (2004) "A Sentimental Education: Sentiment Analysis Using Subjectivity Summarization Based on Minimum Cuts.", Proceedings of the ACL.

Examples

# check polarities
table(data_corpus_moviereviews$sentiment)
#> 
#>  neg  pos 
#> 1000 1000 

# make the data into sentences, because each line is a sentence
data_corpus_moviereviewsents <-
    quanteda::corpus_segment(data_corpus_moviereviews, "\n", extract_pattern = FALSE)
print(data_corpus_moviereviewsents, max_ndoc = 3)
#> Corpus consisting of 64,720 documents and 3 docvars.
#> cv000_29416.txt.1 :
#> "plot : two teen couples go to a church party , drink and the..."
#> 
#> cv000_29416.txt.2 :
#> "they get into an accident ."
#> 
#> cv000_29416.txt.3 :
#> "one of the guys dies , but his girlfriend continues to see h..."
#> 
#> [ reached max_ndoc ... 64,717 more documents ]