Latent Semantic Analysis

Fit the Latent Semantic Analysis scaling model to a dfm, which may be weighted (for instance using quanteda::dfm_tfidf()).

textmodel_lsa(x, nd = 10, margin = c("both", "documents", "features"))

Arguments

x
the dfm on which the model will be fit

nd
the number of dimensions to be included in output

margin
margin to be smoothed by the SVD

Details

svds in the RSpectra package is applied to enable the fast computation of the SVD.

Note

The number of dimensions nd retained in LSA is an empirical issue. While a reduction in \(k\) can remove much of the noise, keeping too few dimensions or factors may lose important information.

References

Rosario, B. (2000). Latent Semantic Indexing: An Overview. Technical report INFOSYS 240 Spring Paper, University of California, Berkeley.

Deerwester, S., Dumais, S.T., Furnas, G.W., Landauer, T.K., & Harshman, R. (1990). Indexing by Latent Semantic Analysis. Journal of the American Society for Information Science, 41(6): 391.

See also

Author

Haiyan Wang and Kohei Watanabe

Examples

dfmat <- quanteda::dfm(data_corpus_irishbudget2010)
# create an LSA space and return its truncated representation in the low-rank space
tmod <- textmodel_lsa(dfmat[1:10, ])

#> Warning: all singular values are requested, svd() is used instead

head(tmod$docs)

#>                            [,1]       [,2]       [,3]        [,4]        [,5]
#> Lenihan, Brian (FF)  -0.5132094  0.6611879  0.5010630  0.03721966 -0.18923845
#> Bruton, Richard (FG) -0.2774009 -0.3444761  0.1537600  0.84967816  0.13612651
#> Burton, Joan (LAB)   -0.3840363 -0.3455336 -0.1080587 -0.22251094 -0.63022740
#> Morgan, Arthur (SF)  -0.4381507 -0.2675656  0.1958102 -0.42933871  0.65837694
#> Cowen, Brian (FF)    -0.3932122  0.3587267 -0.7698225  0.14399779  0.19066465
#> Kenny, Enda (FG)     -0.2611647 -0.1548041 -0.1004087 -0.12286912  0.05830078
#>                              [,6]        [,7]        [,8]        [,9]
#> Lenihan, Brian (FF)   0.024681141 -0.04346863  0.03512717 -0.02562776
#> Bruton, Richard (FG) -0.009252152  0.11188530  0.12487049 -0.10973210
#> Burton, Joan (LAB)    0.021566201  0.51598425  0.04846497 -0.02422325
#> Morgan, Arthur (SF)  -0.206635964  0.15993304  0.10139756  0.01182838
#> Cowen, Brian (FF)    -0.097954789  0.08900744 -0.19260494  0.01589501
#> Kenny, Enda (FG)      0.813893378 -0.37184117  0.08247559 -0.23309870
#>                             [,10]
#> Lenihan, Brian (FF)   0.082441560
#> Bruton, Richard (FG)  0.004660417
#> Burton, Joan (LAB)   -0.071468614
#> Morgan, Arthur (SF)   0.040008687
#> Cowen, Brian (FF)    -0.110079194
#> Kenny, Enda (FG)     -0.132005652

# matrix in low_rank LSA space
tmod$matrix_low_rank[,1:5]

#>                        when  i     presented the supplementary
#> Lenihan, Brian (FF)       5 73  1.000000e+00 539  7.000000e+00
#> Bruton, Richard (FG)      2  6  1.634077e-14 305  1.215135e-13
#> Burton, Joan (LAB)       11 40  9.355364e-15 428  1.732295e-13
#> Morgan, Arthur (SF)      21 26 -2.170745e-12 501  1.000000e+00
#> Cowen, Brian (FF)         4 17 -1.103199e-12 394  6.779993e-14
#> Kenny, Enda (FG)         12 25  1.000000e+00 304  1.000000e+00
#> ODonnell, Kieran (FG)     5 11 -2.283855e-12 193  5.343781e-13
#> Gilmore, Eamon (LAB)      6 10 -2.745295e-12 270 -7.430064e-13
#> Higgins, Michael (LAB)    3  7 -1.262462e-13  78 -3.951284e-13
#> Quinn, Ruairi (LAB)       5 19  1.906982e-13  80  5.900835e-14

# fold queries into the space generated by dfmat[1:10,]
# and return its truncated versions of its representation in the new low-rank space
pred <- predict(tmod, newdata = dfmat[11:14, ])
pred$docs_newspace

#> 4 x 10 Matrix of class "dgeMatrix"
#>                                  [,1]        [,2]        [,3]         [,4]
#> Gormley, John (Green)     -0.06232252  0.02556360  0.01586597  0.002088985
#> Ryan, Eamon (Green)       -0.09764590 -0.05532704 -0.03798698  0.290805183
#> Cuffe, Ciaran (Green)     -0.07289862 -0.01397921 -0.08692541  0.108235208
#> OCaolain, Caoimhghin (SF) -0.24271426 -0.05215588  0.14046794 -0.140695299
#>                                   [,5]         [,6]        [,7]        [,8]
#> Gormley, John (Green)      0.008438226 -0.062441810 -0.01849925 -0.06617083
#> Ryan, Eamon (Green)       -0.059229412 -0.222829324 -0.05353410 -0.01103607
#> Cuffe, Ciaran (Green)      0.031604053 -0.002211501 -0.01642845  0.04111470
#> OCaolain, Caoimhghin (SF)  0.095791137  0.004860009 -0.01768297  0.06001300
#>                                  [,9]       [,10]
#> Gormley, John (Green)      0.01328396 -0.04933352
#> Ryan, Eamon (Green)        0.28532205 -0.19185289
#> Cuffe, Ciaran (Green)      0.07244311 -0.18031934
#> OCaolain, Caoimhghin (SF) -0.07634751  0.23653568