Linear SVM classifier for texts

Fit a fast linear SVM classifier for texts, using the LiblineaR package.

textmodel_svm(x, y, weight = c("uniform", "docfreq", "termfreq"), ...)

Arguments

x
the dfm on which the model will be fit. Does not need to contain only the training documents.

y
vector of training labels associated with each document identified in train. (These will be converted to factors if not already factors.)

weight
weights for different classes for imbalanced training sets, passed to wi in LiblineaR::LiblineaR(). "uniform" uses default; "docfreq" weights by the number of training examples, and "termfreq" by the relative sizes of the training classes in terms of their total lengths in tokens.

...
additional arguments passed to LiblineaR::LiblineaR()

References

R. E. Fan, K. W. Chang, C. J. Hsieh, X. R. Wang, and C. J. Lin. (2008) LIBLINEAR: A Library for Large Linear Classification. Journal of Machine Learning Research 9: 1871-1874. https://www.csie.ntu.edu.tw/~cjlin/liblinear/.

See also

Examples

# use party leaders for govt and opposition classes
quanteda::docvars(data_corpus_irishbudget2010, "govtopp") <-
    c(rep(NA, 4), "Gov", "Opp", NA, "Opp", NA, NA, NA, NA, NA, NA)
dfmat <- quanteda::dfm(data_corpus_irishbudget2010)
tmod <- textmodel_svm(dfmat, y = quanteda::docvars(dfmat, "govtopp"))
predict(tmod)

#>       Lenihan, Brian (FF)      Bruton, Richard (FG)        Burton, Joan (LAB) 
#>                       Gov                       Opp                       Opp 
#>       Morgan, Arthur (SF)         Cowen, Brian (FF)          Kenny, Enda (FG) 
#>                       Opp                       Gov                       Opp 
#>     ODonnell, Kieran (FG)      Gilmore, Eamon (LAB)    Higgins, Michael (LAB) 
#>                       Opp                       Opp                       Opp 
#>       Quinn, Ruairi (LAB)     Gormley, John (Green)       Ryan, Eamon (Green) 
#>                       Opp                       Opp                       Opp 
#>     Cuffe, Ciaran (Green) OCaolain, Caoimhghin (SF) 
#>                       Opp                       Opp 
#> Levels: Gov Opp

predict(tmod, type = "probability")

#>                                    Gov          Opp
#> Lenihan, Brian (FF)       0.5465440857 0.4534559143
#> Bruton, Richard (FG)      0.0009329706 0.9990670294
#> Burton, Joan (LAB)        0.0001561171 0.9998438829
#> Morgan, Arthur (SF)       0.0001134458 0.9998865542
#> Cowen, Brian (FF)         0.9990281614 0.0009718386
#> Kenny, Enda (FG)          0.0010181271 0.9989818729
#> ODonnell, Kieran (FG)     0.0031687655 0.9968312345
#> Gilmore, Eamon (LAB)      0.0008219872 0.9991780128
#> Higgins, Michael (LAB)    0.3131042576 0.6868957424
#> Quinn, Ruairi (LAB)       0.1617482965 0.8382517035
#> Gormley, John (Green)     0.4210807001 0.5789192999
#> Ryan, Eamon (Green)       0.2877136488 0.7122863512
#> Cuffe, Ciaran (Green)     0.4424563288 0.5575436712
#> OCaolain, Caoimhghin (SF) 0.0101010791 0.9898989209

# multiclass problem - all party leaders
tmod2 <- textmodel_svm(dfmat,
    y = c(rep(NA, 3), "SF", "FF", "FG", NA, "LAB", NA, NA, "Green", rep(NA, 3)))
predict(tmod2)

#>       Lenihan, Brian (FF)      Bruton, Richard (FG)        Burton, Joan (LAB) 
#>                     Green                       LAB                       LAB 
#>       Morgan, Arthur (SF)         Cowen, Brian (FF)          Kenny, Enda (FG) 
#>                        SF                        FF                        FG 
#>     ODonnell, Kieran (FG)      Gilmore, Eamon (LAB)    Higgins, Michael (LAB) 
#>                       LAB                       LAB                        FG 
#>       Quinn, Ruairi (LAB)     Gormley, John (Green)       Ryan, Eamon (Green) 
#>                        FG                     Green                     Green 
#>     Cuffe, Ciaran (Green) OCaolain, Caoimhghin (SF) 
#>                     Green                        SF 
#> Levels: FF FG Green LAB SF

predict(tmod2, type = "probability")

#>                                     SF           FF           FG          LAB
#> Lenihan, Brian (FF)       6.897253e-05 8.016667e-05 1.518773e-06 4.605132e-09
#> Bruton, Richard (FG)      1.813693e-01 1.506823e-04 2.009130e-01 5.623238e-01
#> Burton, Joan (LAB)        3.023988e-02 4.305338e-05 1.240717e-01 8.431312e-01
#> Morgan, Arthur (SF)       9.974312e-01 1.006027e-07 1.312651e-03 2.307332e-04
#> Cowen, Brian (FF)         6.339245e-05 9.980543e-01 7.456395e-05 9.151878e-06
#> Kenny, Enda (FG)          9.806327e-04 6.044018e-04 9.953030e-01 2.207272e-03
#> ODonnell, Kieran (FG)     1.774728e-01 1.897339e-04 2.722644e-01 3.789688e-01
#> Gilmore, Eamon (LAB)      5.233783e-04 1.238799e-04 1.147151e-03 9.957988e-01
#> Higgins, Michael (LAB)    5.934149e-02 3.103017e-01 3.625824e-01 2.366221e-01
#> Quinn, Ruairi (LAB)       1.499756e-01 7.220529e-02 3.785460e-01 1.972008e-01
#> Gormley, John (Green)     4.464937e-03 4.935064e-03 8.923454e-03 7.161695e-03
#> Ryan, Eamon (Green)       9.458717e-03 3.052386e-02 6.791777e-03 3.260695e-01
#> Cuffe, Ciaran (Green)     2.975850e-02 1.391622e-01 1.063893e-01 8.437539e-02
#> OCaolain, Caoimhghin (SF) 7.223500e-01 4.151249e-04 6.422164e-02 1.418358e-02
#>                                  Green
#> Lenihan, Brian (FF)       0.9998493374
#> Bruton, Richard (FG)      0.0552431891
#> Burton, Joan (LAB)        0.0025141083
#> Morgan, Arthur (SF)       0.0010252925
#> Cowen, Brian (FF)         0.0017986314
#> Kenny, Enda (FG)          0.0009047057
#> ODonnell, Kieran (FG)     0.1711042699
#> Gilmore, Eamon (LAB)      0.0024068166
#> Higgins, Michael (LAB)    0.0311522723
#> Quinn, Ruairi (LAB)       0.2020723254
#> Gormley, John (Green)     0.9745148505
#> Ryan, Eamon (Green)       0.6271561775
#> Cuffe, Ciaran (Green)     0.6403145935
#> OCaolain, Caoimhghin (SF) 0.1988296503