Wordfish text model

Estimate Slapin and Proksch's (2008) "wordfish" Poisson scaling model of one-dimensional document positions using conditional maximum likelihood.

textmodel_wordfish(
  x,
  dir = c(1, 2),
  priors = c(Inf, Inf, 3, 1),
  tol = c(1e-06, 1e-08),
  dispersion = c("poisson", "quasipoisson"),
  dispersion_level = c("feature", "overall"),
  dispersion_floor = 0,
  abs_err = FALSE,
  residual_floor = 0.5
)

Arguments

x

the dfm on which the model will be fit

dir

set global identification by specifying the indexes for a pair of documents such that \(\hat{\theta}_{dir[1]} < \hat{\theta}_{dir[2]}\).

priors

prior precisions for the estimated parameters \(\alpha_i\), \(\psi_j\), \(\beta_j\), and \(\theta_i\), where \(i\) indexes documents and \(j\) indexes features

tol

tolerances for convergence. The first value is a convergence threshold for the log-posterior of the model, the second value is the tolerance in the difference in parameter values from the iterative conditional maximum likelihood (from conditionally estimating document-level, then feature-level parameters).

dispersion

sets whether a quasi-Poisson quasi-likelihood should be used based on a single dispersion parameter ("poisson"), or quasi-Poisson ("quasipoisson")

dispersion_level

sets the unit level for the dispersion parameter, options are "feature" for term-level variances, or "overall" for a single dispersion parameter

dispersion_floor

constraint for the minimal underdispersion multiplier in the quasi-Poisson model. Used to minimize the distorting effect of terms with rare term or document frequencies that appear to be severely underdispersed. Default is 0, but this only applies if dispersion = "quasipoisson".

abs_err

specifies how the convergence is considered

residual_floor

specifies the threshold for residual matrix when calculating the svds, only applies when sparse = TRUE

Value

An object of class textmodel_fitted_wordfish. This is a list containing:

dir

global identification of the dimension

theta

estimated document positions

alpha

estimated document fixed effects

beta

estimated feature marginal effects

psi

estimated word fixed effects

docs

document labels

features

feature labels

sigma

regularization parameter for betas in Poisson form

ll

log likelihood at convergence

se.theta

standard errors for theta-hats

x

dfm to which the model was fit

Details

The returns match those of Will Lowe's R implementation of wordfish (see the austin package), except that here we have renamed words to be features. (This return list may change.) We have also followed the practice begun with Slapin and Proksch's early implementation of the model that used a regularization parameter of se\((\sigma) = 3\), through the third element in priors.

Note

In the rare situation where a warning message of "The algorithm did not converge." shows up, removing some documents may work.

References

Slapin, J. & Proksch, S.O. (2008). A Scaling Model for Estimating Time-Series Party Positions from Texts. doi:10.1111/j.1540-5907.2008.00338.x . American Journal of Political Science, 52(3), 705--772.

Lowe, W. & Benoit, K.R. (2013). Validating Estimates of Latent Traits from Textual Data Using Human Judgment as a Benchmark. doi:10.1093/pan/mpt002 . Political Analysis, 21(3), 298--313.

See also

Author

Benjamin Lauderdale, Haiyan Wang, and Kenneth Benoit

Examples

(tmod1 <- textmodel_wordfish(quanteda::data_dfm_lbgexample, dir = c(1,5)))
#> 
#> Call:
#> textmodel_wordfish.dfm(x = quanteda::data_dfm_lbgexample, dir = c(1, 
#>     5))
#> 
#> Dispersion: poisson; direction: 1 < 5; 6 documents; 37 features.
summary(tmod1, n = 10)
#> 
#> Call:
#> textmodel_wordfish.dfm(x = quanteda::data_dfm_lbgexample, dir = c(1, 
#>     5))
#> 
#> Estimated Document Positions:
#>       theta      se
#> R1 -1.29057 0.01448
#> R2 -0.61644 0.01060
#> R3  0.03351 0.01085
#> R4  0.70961 0.01214
#> R5  1.52334 0.01744
#> V1 -0.35945 0.01070
#> 
#> Estimated Feature Scores:
#>           A      B       C       D      E       F      G       H       I
#> beta -8.445 -9.300 -11.258 -12.364 -13.33 -13.237 -13.49 -12.359 -11.585
#> psi  -6.641 -6.989  -7.869  -8.404  -8.89  -8.196  -8.12  -6.426  -5.345
#>            J
#> beta -10.424
#> psi   -3.928
coef(tmod1)
#> $documents
#>          theta     alpha
#> R1 -1.29056981 -4.558205
#> R2 -0.61643907  1.303010
#> R3  0.03350982  3.307253
#> R4  0.70961244  1.688789
#> R5  1.52333687 -4.270413
#> V1 -0.35945025  2.529567
#> 
#> $features
#>           beta        psi
#> A   -8.4452617 -6.6414630
#> B   -9.3002556 -6.9890837
#> C  -11.2580501 -7.8685425
#> D  -12.3643578 -8.4041920
#> E  -13.3323903 -8.8901495
#> F  -13.2373464 -8.1956515
#> G  -13.4867108 -8.1199814
#> H  -12.3588189 -6.4258549
#> I  -11.5848271 -5.3452709
#> J  -10.4239011 -3.9281122
#> K   -9.2561345 -2.6663316
#> L   -8.1485428 -1.6133457
#> M   -6.9820863 -0.6511295
#> N   -5.8782223  0.1109961
#> O   -4.7955288  0.7252203
#> P   -3.5739210  1.2419336
#> Q   -2.5522985  1.5366923
#> R   -1.3458753  1.7340400
#> S   -0.2889734  1.7502790
#> T    0.7696623  1.6318538
#> U    1.9353486  1.3496139
#> V    2.9295159  0.9736353
#> W    4.1021898  0.3716688
#> X    5.0019413 -0.2132110
#> Y    5.9373363 -0.9619285
#> Z    6.8722030 -1.8640402
#> ZA   7.7519960 -2.8378397
#> ZB   8.7447725 -4.1062621
#> ZC   9.6708363 -5.4394764
#> ZD  10.4464336 -6.7008977
#> ZE  11.2825127 -8.2159091
#> ZF  11.1103336 -8.3482660
#> ZG  11.4115094 -9.3724297
#> ZH  10.5822264 -8.8587331
#> ZI   9.6390952 -8.2849022
#> ZJ   8.0273986 -7.3432860
#> ZK   7.3723185 -6.9825744
#> 
predict(tmod1)
#>          R1          R2          R3          R4          R5          V1 
#> -1.29056981 -0.61643907  0.03350982  0.70961244  1.52333687 -0.35945025 
predict(tmod1, se.fit = TRUE)
#> $fit
#>          R1          R2          R3          R4          R5          V1 
#> -1.29056981 -0.61643907  0.03350982  0.70961244  1.52333687 -0.35945025 
#> 
#> $se.fit
#> [1] 0.01448422 0.01059613 0.01085375 0.01214106 0.01744178 0.01070073
#> 
predict(tmod1, interval = "confidence")
#> $fit
#>            fit         lwr         upr
#> R1 -1.29056981 -1.31895837 -1.26218126
#> R2 -0.61643907 -0.63720710 -0.59567103
#> R3  0.03350982  0.01223686  0.05478279
#> R4  0.70961244  0.68581641  0.73340848
#> R5  1.52333687  1.48915160  1.55752213
#> V1 -0.35945025 -0.38042330 -0.33847721
#> 

if (FALSE) {
library("quanteda")
dfmat <- dfm(tokens(data_corpus_irishbudget2010))
(tmod2 <- textmodel_wordfish(dfmat, dir = c(6,5)))
(tmod3 <- textmodel_wordfish(dfmat, dir = c(6,5),
                             dispersion = "quasipoisson", dispersion_floor = 0))
(tmod4 <- textmodel_wordfish(dfmat, dir = c(6,5),
                             dispersion = "quasipoisson", dispersion_floor = .5))
plot(tmod3$phi, tmod4$phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
     xlim = c(0, 1.0), ylim = c(0, 1.0))
plot(tmod3$phi, tmod4$phi, xlab = "Min underdispersion = 0", ylab = "Min underdispersion = .5",
     xlim = c(0, 1.0), ylim = c(0, 1.0), type = "n")
underdispersedTerms <- sample(which(tmod3$phi < 1.0), 5)
which(featnames(dfmat) %in% names(topfeatures(dfmat, 20)))
text(tmod3$phi, tmod4$phi, tmod3$features,
     cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "grey90")
text(tmod3$phi['underdispersedTerms'], tmod4$phi['underdispersedTerms'],
     tmod3$features['underdispersedTerms'],
     cex = .8, xlim = c(0, 1.0), ylim = c(0, 1.0), col = "black")
if (requireNamespace("austin")) {
    tmod5 <- austin::wordfish(quanteda::as.wfm(dfmat), dir = c(6, 5))
    cor(tmod1$theta, tmod5$theta)
}}