quanteda.textmodels: Scaling Models and Classifiers for Textual Data

Scaling models and classifiers for sparse matrix objects representing textual data in the form of a document-feature matrix. Includes original implementations of 'Laver', 'Benoit', and Garry's (2003) doi:10.1017/S0003055403000698 , 'Wordscores' model, the Perry and 'Benoit' (2017) arXiv:1710.08963 class affinity scaling model, and the 'Slapin' and 'Proksch' (2008) doi:10.1111/j.1540-5907.2008.00338.x 'wordfish' model, as well as methods for correspondence analysis, latent semantic analysis, and fast Naive Bayes and linear 'SVMs' specially designed for sparse textual data.

See also

Author

Maintainer: Kenneth Benoit kbenoit@lse.ac.uk (ORCID) [copyright holder]

Authors:

Kohei Watanabe watanabe.kohei@gmail.com (ORCID)

Haiyan Wang whyinsa@yahoo.com (ORCID)

Patrick O. Perry patperry@gmail.com (ORCID)

Benjamin Lauderdale b.e.lauderdale@lse.ac.uk (ORCID)

Johannes Gruber JohannesB.Gruber@gmail.com (ORCID)

William Lowe lowe@hertie-school.org (ORCID)

Other contributors:

Vikas Sindhwani vikas.sindhwani@gmail.com (authored svmlin C++ source code) [copyright holder]

European Research Council (ERC-2011-StG 283794-QUANTESS) [funder]