Logistic regression classifier for texts

Fits a fast penalized maximum likelihood estimator to predict discrete categories from sparse dfm objects. Using the glmnet package, the function computes the regularization path for the lasso or elasticnet penalty at a grid of values for the regularization parameter lambda. This is done automatically by testing on several folds of the data at estimation time.

textmodel_lr(x, y, ...)

Arguments

x

the dfm on which the model will be fit. Does not need to contain only the training documents.

y

vector of training labels associated with each document identified in train. (These will be converted to factors if not already factors.)

...

additional arguments passed to cv.glmnet()

Value

an object of class textmodel_lr, a list containing:

x, y the input model matrix and input training class labels

algorithm character; the type and family of logistic regression model used in calling cv.glmnet()

type the type of associated with algorithm

classnames the levels of training classes in y

lrfitted the fitted model object from cv.glmnet()

call the model call

References

Friedman, J., Hastie, T., & Tibshirani, R. (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software 33(1), 1-22. doi:10.18637/jss.v033.i01

See also

Examples

## Example from 13.1 of _An Introduction to Information Retrieval_
library("quanteda")
corp <- corpus(c(d1 = "Chinese Beijing Chinese",
                 d2 = "Chinese Chinese Shanghai",
                 d3 = "Chinese Macao",
                 d4 = "Tokyo Japan Chinese",
                 d5 = "London England Chinese",
                 d6 = "Chinese Chinese Chinese Tokyo Japan"),
               docvars = data.frame(train = factor(c("Y", "Y", "Y", "N", "N", NA))))
dfmat <- dfm(tokens(corp), tolower = FALSE)

## simulate bigger sample as classification on small samples is problematic
set.seed(1)
dfmat <- dfm_sample(dfmat, 50, replace = TRUE)

## train model
(tmod1 <- textmodel_lr(dfmat, docvars(dfmat, "train")))
#> 
#> Call:
#> textmodel_lr.dfm(x = dfmat, y = docvars(dfmat, "train"))
#> 
#> 50 training documents; 8 fitted features.
#> Method: binomial logistic regression
summary(tmod1)
#> 
#> Call:
#> textmodel_lr.dfm(x = dfmat, y = docvars(dfmat, "train"))
#> 
#> Lambda Min:
#> [1] 0.0004211
#> 
#> Lambda 1se:
#> [1] 0.0004211
#> 
#> Estimated Feature Scores:
#>   (Intercept) Chinese Beijing Shanghai Macao  Tokyo      Japan London
#> Y       6.115  0.8428       0        0     0 -14.02 -9.635e-15 -13.94
#>      England
#> Y -2.754e-14
coef(tmod1)
#> 9 x 1 sparse Matrix of class "dgCMatrix"
#>                         Y
#> (Intercept)  6.114526e+00
#> Chinese      8.427810e-01
#> Beijing      .           
#> Shanghai     .           
#> Macao        .           
#> Tokyo       -1.401864e+01
#> Japan       -9.635062e-15
#> London      -1.393844e+01
#> England     -2.753528e-14

## predict probability and classes
predict(tmod1, type = "prob")
#>                  Y            N
#> d1.1  0.9995904688 0.0004095312
#> d4.1  0.0008569013 0.9991430987
#> d1.2  0.9995904688 0.0004095312
#> d2.1  0.9995904688 0.0004095312
#> d5.1  0.0009283860 0.9990716140
#> d3.1  0.9990492486 0.0009507514
#> d6.1  0.0046060476 0.9953939524
#> d2.2  0.9995904688 0.0004095312
#> d3.2  0.9990492486 0.0009507514
#> d3.3  0.9990492486 0.0009507514
#> d1.3  0.9995904688 0.0004095312
#> d5.2  0.0009283860 0.9990716140
#> d5.3  0.0009283860 0.9990716140
#> d2.3  0.9995904688 0.0004095312
#> d6.2  0.0046060476 0.9953939524
#> d6.3  0.0046060476 0.9953939524
#> d2.4  0.9995904688 0.0004095312
#> d1.4  0.9995904688 0.0004095312
#> d5.4  0.0009283860 0.9990716140
#> d5.5  0.0009283860 0.9990716140
#> d1.5  0.9995904688 0.0004095312
#> d1.6  0.9995904688 0.0004095312
#> d6.4  0.0046060476 0.9953939524
#> d5.6  0.0009283860 0.9990716140
#> d5.7  0.0009283860 0.9990716140
#> d2.5  0.9995904688 0.0004095312
#> d2.6  0.9995904688 0.0004095312
#> d6.5  0.0046060476 0.9953939524
#> d1.7  0.9995904688 0.0004095312
#> d4.2  0.0008569013 0.9991430987
#> d1.8  0.9995904688 0.0004095312
#> d4.3  0.0008569013 0.9991430987
#> d3.4  0.9990492486 0.0009507514
#> d6.6  0.0046060476 0.9953939524
#> d2.7  0.9995904688 0.0004095312
#> d2.8  0.9995904688 0.0004095312
#> d6.7  0.0046060476 0.9953939524
#> d4.4  0.0008569013 0.9991430987
#> d4.5  0.0008569013 0.9991430987
#> d4.6  0.0008569013 0.9991430987
#> d2.9  0.9995904688 0.0004095312
#> d4.7  0.0008569013 0.9991430987
#> d1.9  0.9995904688 0.0004095312
#> d6.8  0.0046060476 0.9953939524
#> d1.10 0.9995904688 0.0004095312
#> d4.8  0.0008569013 0.9991430987
#> d1.11 0.9995904688 0.0004095312
#> d6.9  0.0046060476 0.9953939524
#> d2.10 0.9995904688 0.0004095312
#> d3.5  0.9990492486 0.0009507514
predict(tmod1)
#>  d1.1  d4.1  d1.2  d2.1  d5.1  d3.1  d6.1  d2.2  d3.2  d3.3  d1.3  d5.2  d5.3 
#>     Y     N     Y     Y     N     Y     N     Y     Y     Y     Y     N     N 
#>  d2.3  d6.2  d6.3  d2.4  d1.4  d5.4  d5.5  d1.5  d1.6  d6.4  d5.6  d5.7  d2.5 
#>     Y     N     N     Y     Y     N     N     Y     Y     N     N     N     Y 
#>  d2.6  d6.5  d1.7  d4.2  d1.8  d4.3  d3.4  d6.6  d2.7  d2.8  d6.7  d4.4  d4.5 
#>     Y     N     Y     N     Y     N     Y     N     Y     Y     N     N     N 
#>  d4.6  d2.9  d4.7  d1.9  d6.8 d1.10  d4.8 d1.11  d6.9 d2.10  d3.5 
#>     N     Y     N     Y     N     Y     N     Y     N     Y     Y 
#> Levels: N Y